requirements.txt
requests~=2.32.3
playwright~=1.46.0
qianfan~=0.4.8
python-dotenv~=1.0.1
gradio~=4.44.0